requirements.txt
# AWS S3 and PDF processing (read_files.py)
boto3>=1.35.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0

# Vector embeddings and ML (vector.py, chunk.py)
sentence-transformers>=3.0.0
torch>=2.0.0
numpy>=1.24.0

# Scientific computing (chunk.py - find_peaks)
scipy>=1.11.0

# Database (db_writer.py)
psycopg2-binary>=2.9.0
# Note: Requires PostgreSQL with pgvector extension installed

# RAG System (gemini.py)
google-genai>=1.56.0

